# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is the api for generated Notion forms",
        "title": "Generated Notion Forms",
        "contact": {
            "name": "API Support",
            "email": "develop@generated-notion-forms.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/notion/database": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return a list of all own databases, where you have given access to it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notion"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.MinimalistDatabase"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    }
                }
            }
        },
        "/notion/database/{databaseId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return a own databases, where you have given access to it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notion"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "databaseId",
                        "name": "databaseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notionapi.Database"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return a own databases, where you have given access to it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notion"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "databaseId",
                        "name": "databaseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "RecordRequest",
                        "name": "RecordRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.RecordRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notionapi.Page"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    }
                }
            }
        },
        "/notion/database/{databaseId}/properties": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return the properties of a database by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notion"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "databaseId",
                        "name": "databaseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DatabasePropertyResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    }
                }
            }
        },
        "/notion/database/{databaseId}/properties/options": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return a list of all select options from a database by their id, where you have given access to it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notion"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "databaseId",
                        "name": "databaseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Select"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    }
                }
            }
        },
        "/notion/database/{databaseId}/properties/options/{notionSelectId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return a list of all select options from a database by their id, where you have given access to it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notion"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "databaseId",
                        "name": "databaseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "notionSelectId",
                        "name": "notionSelectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Select"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.HttpError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.DatabasePropertyResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "model.HttpError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.MinimalistDatabase": {
            "type": "object",
            "properties": {
                "created_time": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.Option": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.RecordRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "model.Select": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Option"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "notionapi.Annotations": {
            "type": "object",
            "properties": {
                "bold": {
                    "type": "boolean"
                },
                "code": {
                    "type": "boolean"
                },
                "color": {
                    "$ref": "#/definitions/notionapi.Color"
                },
                "italic": {
                    "type": "boolean"
                },
                "strikethrough": {
                    "type": "boolean"
                },
                "underline": {
                    "type": "boolean"
                }
            }
        },
        "notionapi.Bot": {
            "type": "object",
            "properties": {
                "owner": {
                    "$ref": "#/definitions/notionapi.Owner"
                },
                "workspace_name": {
                    "type": "string"
                }
            }
        },
...
        "notionapi.Database": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "cover": {
                    "$ref": "#/definitions/notionapi.Image"
                },
                "created_by": {
                    "$ref": "#/definitions/notionapi.User"
                },
                "created_time": {
                    "type": "string"
                },
                "description": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/notionapi.RichText"
                    }
                },
                "icon": {
                    "$ref": "#/definitions/notionapi.Icon"
                },
                "id": {
                    "type": "string"
                },
                "is_inline": {
                    "type": "boolean"
                },
                "last_edited_by": {
                    "$ref": "#/definitions/notionapi.User"
                },
                "last_edited_time": {
                    "type": "string"
                },
                "object": {
                    "$ref": "#/definitions/notionapi.ObjectType"
                },
                "parent": {
                    "$ref": "#/definitions/notionapi.Parent"
                },
                "properties": {
                    "description": "Properties is a map of property configurations that defines what Page.Properties each page of the database can use",
                    "allOf": [
                        {
                            "$ref": "#/definitions/notionapi.PropertyConfigs"
                        }
                    ]
                },
                "title": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/notionapi.RichText"
                    }
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "notionapi.DatabaseMention": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "notionapi.DateObject": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "notionapi.Equation": {
            "type": "object",
            "properties": {
                "expression": {
                    "type": "string"
                }
            }
        },
        "notionapi.FileObject": {
            "type": "object",
            "properties": {
                "expiry_time": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "notionapi.FileType": {
            "type": "string",
            "enum": [
                "file",
                "external"
            ],
            "x-enum-varnames": [
                "FileTypeFile",
                "FileTypeExternal"
            ]
        },
        "notionapi.Icon": {
            "type": "object",
            "properties": {
                "emoji": {
                    "type": "string"
                },
                "external": {
                    "$ref": "#/definitions/notionapi.FileObject"
                },
                "file": {
                    "$ref": "#/definitions/notionapi.FileObject"
                },
                "type": {
                    "$ref": "#/definitions/notionapi.FileType"
                }
            }
        },
        "notionapi.Image": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/notionapi.RichText"
                    }
                },
                "external": {
                    "$ref": "#/definitions/notionapi.FileObject"
                },
                "file": {
                    "$ref": "#/definitions/notionapi.FileObject"
                },
                "type": {
                    "$ref": "#/definitions/notionapi.FileType"
                }
            }
        },
        "notionapi.Link": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "notionapi.Mention": {
            "type": "object",
            "properties": {
                "database": {
                    "$ref": "#/definitions/notionapi.DatabaseMention"
                },
                "date": {
                    "$ref": "#/definitions/notionapi.DateObject"
                },
                "page": {
                    "$ref": "#/definitions/notionapi.PageMention"
                },
                "template_mention": {
                    "$ref": "#/definitions/notionapi.TemplateMention"
                },
                "type": {
                    "$ref": "#/definitions/notionapi.MentionType"
                },
                "user": {
                    "$ref": "#/definitions/notionapi.User"
                }
            }
        },
...
        "notionapi.ObjectType": {
            "type": "string",
            "enum": [
                "database",
                "block",
                "page",
                "list",
                "text",
                "user",
                "error",
                "comment"
            ],
            "x-enum-varnames": [
                "ObjectTypeDatabase",
                "ObjectTypeBlock",
                "ObjectTypePage",
                "ObjectTypeList",
                "ObjectTypeText",
                "ObjectTypeUser",
                "ObjectTypeError",
                "ObjectTypeComment"
            ]
        },
        "notionapi.Owner": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "workspace": {
                    "type": "boolean"
                }
            }
        },
        "notionapi.Page": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "cover": {
                    "$ref": "#/definitions/notionapi.Image"
                },
                "created_by": {
                    "$ref": "#/definitions/notionapi.User"
                },
                "created_time": {
                    "type": "string"
                },
                "icon": {
                    "$ref": "#/definitions/notionapi.Icon"
                },
                "id": {
                    "type": "string"
                },
                "last_edited_by": {
                    "$ref": "#/definitions/notionapi.User"
                },
                "last_edited_time": {
                    "type": "string"
                },
                "object": {
                    "$ref": "#/definitions/notionapi.ObjectType"
                },
                "parent": {
                    "$ref": "#/definitions/notionapi.Parent"
                },
                "properties": {
                    "$ref": "#/definitions/notionapi.Properties"
                },
                "public_url": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "notionapi.PageMention": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "notionapi.Parent": {
            "type": "object",
            "properties": {
                "block_id": {
                    "type": "string"
                },
                "database_id": {
                    "type": "string"
                },
                "page_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/notionapi.ParentType"
                },
                "workspace": {
                    "type": "boolean"
                }
            }
        },
        "notionapi.ParentType": {
            "type": "string",
            "enum": [
                "database_id",
                "page_id",
                "workspace",
                "block_id"
            ],
            "x-enum-varnames": [
                "ParentTypeDatabaseID",
                "ParentTypePageID",
                "ParentTypeWorkspace",
                "ParentTypeBlockID"
            ]
        },
        "notionapi.Person": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "notionapi.Properties": {
            "type": "object",
            "additionalProperties": {}
        },
        "notionapi.PropertyConfigs": {
            "type": "object",
            "additionalProperties": {}
        },
        "notionapi.RichText": {
            "type": "object",
            "properties": {
                "annotations": {
                    "$ref": "#/definitions/notionapi.Annotations"
                },
                "equation": {
                    "$ref": "#/definitions/notionapi.Equation"
                },
                "href": {
                    "type": "string"
                },
                "mention": {
                    "$ref": "#/definitions/notionapi.Mention"
                },
                "plain_text": {
                    "type": "string"
                },
                "text": {
                    "$ref": "#/definitions/notionapi.Text"
                },
                "type": {
                    "$ref": "#/definitions/notionapi.ObjectType"
                }
            }
        },
...
        "notionapi.Text": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "link": {
                    "$ref": "#/definitions/notionapi.Link"
                }
            }
        },
        "notionapi.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bot": {
                    "$ref": "#/definitions/notionapi.Bot"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "object": {
                    "$ref": "#/definitions/notionapi.ObjectType"
                },
                "person": {
                    "$ref": "#/definitions/notionapi.Person"
                },
                "type": {
                    "$ref": "#/definitions/notionapi.UserType"
                }
            }
        },
        "notionapi.UserType": {
            "type": "string",
            "enum": [
                "person",
                "bot"
            ],
            "x-enum-varnames": [
                "UserTypePerson",
                "UserTypeBot"
            ]
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}